Enable #nullable for Group/Join family of operators.
