+++ Rx.NET/Source/src/System.Reactive/Linq/Observable/GroupBy.cs
@@ -2,8 +2,6 @@
 // The .NET Foundation licenses this file to you under the MIT License.
 // See the LICENSE file in the project root for more information. 
 
-#nullable disable
-
 using System.Collections.Generic;
 using System.Reactive.Disposables;
 using System.Reactive.Subjects;
@@ -35,10 +33,10 @@ namespace System.Reactive.Linq.ObservableImpl
         {
             private readonly Func<TSource, TKey> _keySelector;
             private readonly Func<TSource, TElement> _elementSelector;
-            private readonly Dictionary<TKey, Subject<TElement>> _map;
+            private readonly Grouping<TKey, TElement> _map;
 
-            private RefCountDisposable _refCountDisposable;
-            private Subject<TElement> _null;
+            private RefCountDisposable? _refCountDisposable;
+            private Subject<TElement>? _null;
 
             public _(GroupBy<TSource, TKey, TElement> parent, IObserver<IGroupedObservable<TKey, TElement>> observer)
                 : base(observer)
@@ -48,11 +46,11 @@ namespace System.Reactive.Linq.ObservableImpl
 
                 if (parent._capacity.HasValue)
                 {
-                    _map = new Dictionary<TKey, Subject<TElement>>(parent._capacity.Value, parent._comparer);
+                    _map = new Grouping<TKey, TElement>(parent._capacity.Value, parent._comparer);
                 }
                 else
                 {
-                    _map = new Dictionary<TKey, Subject<TElement>>(parent._comparer);
+                    _map = new Grouping<TKey, TElement>(parent._comparer);
                 }
             }
 
@@ -79,7 +77,7 @@ namespace System.Reactive.Linq.ObservableImpl
                 }
 
                 var fireNewMapEntry = false;
-                Subject<TElement> writer;
+                Subject<TElement>? writer;
                 try
                 {
                     //
@@ -128,7 +126,7 @@ namespace System.Reactive.Linq.ObservableImpl
 
                 if (fireNewMapEntry)
                 {
-                    var group = new GroupedObservable<TKey, TElement>(key, writer, _refCountDisposable);
+                    var group = new GroupedObservable<TKey, TElement>(key, writer, _refCountDisposable!); // NB: _refCountDisposable is set in Run.
                     ForwardOnNext(group);
                 }
